added saving of bot message to dialogs psql table

+++ src/connect_to_forum/main.py
@@ -6,7 +6,7 @@ import re
 import urllib.parse
 from time import sleep
 from bs4 import BeautifulSoup
-from telegram import ReplyKeyboardMarkup, ForceReply, KeyboardButton, Bot, Update
+from telegram import ReplyKeyboardMarkup, Bot
 import psycopg2
 import datetime
 import logging
@@ -202,9 +202,6 @@ def main(event, context):
     data_in_ascii = encoded_to_ascii['data']
     message_in_ascii = data_in_ascii['message']
     tg_user_id, f_username = list(message_in_ascii)
-    print('DBG.F.2.:Pub/sub:', message_in_ascii)
-    # print(str(tg_user_id), type(tg_user_id))
-    # print(f_username)
 
     # initiate Prod Bot
     bot_token = get_secrets("bot_api_token__prod")
@@ -227,43 +224,51 @@ def main(event, context):
                 get_user_data(block_of_user_data)
 
     if user_found:
-        msg = 'Посмотрите, Бот нашел следующий аккаунт на форуме, это Вы?\n'
-        msg += 'username: ' + f_username + ', '
+        bot_message = 'Посмотрите, Бот нашел следующий аккаунт на форуме, это Вы?\n'
+        bot_message += 'username: ' + f_username + ', '
         if user.callsign:
-            msg += 'позывной: ' + user.callsign + ', '
+            bot_message += 'позывной: ' + user.callsign + ', '
         if user.region:
-            msg += 'регион: ' + user.region + ', '
+            bot_message += 'регион: ' + user.region + ', '
         if user.phone:
-            msg += 'телефон оканчивается на ' + str(user.phone)[-5:] + ', '
+            bot_message += 'телефон оканчивается на ' + str(user.phone)[-5:] + ', '
         if user.age:
-            msg += 'возраст: ' + str(user.age) + ', '
+            bot_message += 'возраст: ' + str(user.age) + ', '
         if user.reg_date:
-            msg += 'дата регистрации: ' + str(user.reg_date)[:-7] + ', '
-        msg = msg[:-2]
+            bot_message += 'дата регистрации: ' + str(user.reg_date)[:-7] + ', '
+        bot_message = bot_message[:-2]
 
         keyboard = [['да, это я'], ['нет, это не я'], ['в начало']]
 
         sql_connect_by_psycopg2()
 
         # Delete previous records for this user
-        cur.execute("""DELETE FROM user_forum_attributes WHERE
-        user_id=%s;""", (tg_user_id,))
+        cur.execute("""DELETE FROM user_forum_attributes WHERE user_id=%s;""", (tg_user_id,))
         conn_psy.commit()
 
         # Add new record for this user
         cur.execute("""INSERT INTO user_forum_attributes 
-        (user_id, forum_user_id, status, timestamp, forum_username, forum_age, forum_sex, forum_region, forum_auto_num, forum_callsign, forum_phone, forum_reg_date) 
+        (user_id, forum_user_id, status, timestamp, forum_username, forum_age, forum_sex, forum_region, 
+        forum_auto_num, forum_callsign, forum_phone, forum_reg_date) 
         values (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);""",
                     (tg_user_id, f_usr_id, 'non-varified', datetime.datetime.now(), f_username, user.age, user.sex,
                      user.region, user.auto_num, user.callsign, user.phone, user.reg_date))
         conn_psy.commit()
 
     else:
-        msg = 'Бот не смог найти такого пользователя на форуме. Пожалуйста, проверьте правильность написания имени пользвателя (логина). Важно, чтобы каждый знак в точности соответсовал тому, что указано в вашем профиле на форуме'
-        keyboard = [['связать аккаунты бота и форума'], ['в начало']]
+        bot_message = 'Бот не смог найти такого пользователя на форуме. ' \
+                      'Пожалуйста, проверьте правильность написания имени пользвателя (логина). ' \
+                      'Важно, чтобы каждый знак в точности соответсовал тому, что указано в вашем профиле на форуме'
+        keyboard = [['в начало']]
 
     reply_markup = ReplyKeyboardMarkup(keyboard, resize_keyboard=True)
-    bot.sendMessage(chat_id=tg_user_id, text=msg, reply_markup=reply_markup, parse_mode='HTML')
+    bot.sendMessage(chat_id=tg_user_id, text=bot_message, reply_markup=reply_markup, parse_mode='HTML')
+
+    # save bot's reply to incoming request
+    if bot_message:
+        cur.execute("""INSERT INTO dialogs (user_id, author, timestamp, message_text) values (%s, %s, %s, %s);""",
+                    (tg_user_id, 'bot', datetime.datetime.now(), bot_message))
+        conn_psy.commit()
 
     cur.close()
     conn_psy.close()
